Move SDC fileutil to libcommon
